+++ vlib/term/term.v
@@ -7,10 +7,11 @@ const (
 	default_rows_size = 25
 )
 
+// Coord - used by term.get_cursor_position and term.set_cursor_position
 pub struct Coord {
 pub:
-	x i16
-	y i16
+	x int
+	y int
 }
 
 // can_show_color_on_stdout returns true if colors are allowed in stdout;
